requirements.txt
bottle
requests
pyserial
